+++ TowerObjects/TowerObjects/StrongInvader.cs
@@ -5,7 +5,7 @@ namespace TowerDefense
 {
     class StrongInvader : Invader
     {
-        public override int Health { get; protected set; } = 5;
+        public override double Health { get; protected set; } = 5;
 
         public StrongInvader(Path path) :base(path)
         {
@@ -13,7 +13,7 @@ namespace TowerDefense
         }
 
 
-        public override void DecreaseHealth(int factor) 
+        public override void DecreaseHealth(double factor) 
         {
             Health -= factor;
             if (IsNeutralized) { Console.WriteLine("GAHHHHHH!!!"); }
